2.0.0 Core Classes
